Create Animated GIF from the Raster Time Series

This function takes the entire raster time series and makes an animated GIF.

Usage

Animation(
  raster_dir = NULL,
  raster_type = "tif",
  color_palette = "BluYl",
  range = c(370, 380),
  magick_path = "C:/Program Files/ImageMagick-7.0.10-Q16-HDRI",
  out_file = "animation.gif",
  ...
)

Arguments

raster_dir

Directory, where the rasters for the animation are stored. Filenames are assumed to include the timestamp (see details)

raster_type

File type of the input rasters stored in raster_dir. Default format is tif

color_palette

Color palette to use for plotting the single raster files. Any palette from the colorspace-package is supported.

magick_path

Parent directory of ImageMagick. Can be downloaded from https://imagemagick.org/

out_file

Path to the resulting GIF file

...

Further parameters for ani.options and saveGIF. See help("ani.options") and help("saveGIF") for further information.

Value

A GIF file with the animation of the raster time series as specified with 'out_file'

Details

The functions assumes to take the output from the predict_rts() function from the GroundwatermodelR package where file names include the timestamp: gwl_YYYYmmdd (gwd refers to groundwater level). For instance gwl_20201025.tif refers to the interpolation raster on 2020-10-25. An option to include custom filename formats is not included yet.
The function assumes 'magick.exe' to be present. See help("ani.options") for further details.

References

Betz, F; Fischer, P. (in prep.): Assessing Spatial-Temporal Dynamics of Groundwater in a Restored Floodplain at the German Danube using Machine Learning. In Preparation.

Author

Florian Betz